moved device widget modifying methods into a new DeviceManager class
