+++ src/managers/__init__.py
@@ -1,2 +1,3 @@
 from .data_holder import DataHolder
 from .timer_service import TimerService
+from .device_manager import DeviceManager
